+++ plugins/modules/cp_mgmt_access_rule.py
@@ -326,7 +326,6 @@ cp_mgmt_access_rule:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.connection import Connection
 from ansible_collections.check_point.mgmt.plugins.module_utils.checkpoint import (
     checkpoint_argument_spec_for_objects,
     api_call,